Check for single node sts:AssumeRole cycles
When a role attempts to assume it self an explicit allow in the
trust policy is not strictly necessary for access. An explicit
allow in the identity policy is sufficient.

This applies to the circular access check because role revocations
do not effect very recently refreshed credentials. When a role can
assume it self, a bash loop calling sts assume-role and exporting
the new credentials to the environment will bypass a revocation
triggered through the web UI.